[ci] rework e2e integration trigger by using labels versus on-hold jobs (#1449)
